+++ test/adapter/json_api/collection_test.rb
@@ -19,7 +19,7 @@ module ActiveModel
             @second_post.author = @author
             @author.posts = [@first_post, @second_post]
 
-            @serializer = ArraySerializer.new([@first_post, @second_post])
+            @serializer = CollectionSerializer.new([@first_post, @second_post])
             @adapter = ActiveModel::Serializer::Adapter::JsonApi.new(@serializer)
             ActionController::Base.cache_store.clear
           end